services/essentia-analyzer/requirements.txt
essentia-tensorflow==2.1b6.dev1110
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
numpy==1.26.4
